Version 2.4.0: Add support for 64-bit doubles.
This program now compiles and runs correctly under 2.11BSD:

```

int main() {
  double x = 12;
  printf("x is %f\n", x);
  return(0);
}
```

+++ defines.h
@@ -128,7 +128,7 @@ typedef unsigned long u_int32_t;
 typedef void (*_itab) P((void));
 extern _itab itab[];		/* Instruction Table for Fast Decode. */
 
-typedef float FLOAT;		/* For now, we use floats to do FP */
+typedef double FLOAT;		/* We now use floats to do FP */
 
 /* PDP processor defines. */
